backend/requirements.txt
# Core ML & Data Processing
torch>=1.13.0
numpy>=1.23.0
pandas>=1.5.0
sentence-transformers>=2.2.0
scikit-learn>=1.2.0

# Image Processing
pillow>=9.0.0

# YouTube & Web APIs
requests>=2.28.0
youtube-transcript-api>=0.6.0

# Web Framework (Backend)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Frontend (Optional)
streamlit>=1.20.0
